Restructure library to be more consistent with existing solution
