--- Test/Check/Function.hs
@@ -1,2 +1,0 @@
-module Test.Check.Function () where
-import Test.Check.Function.ListsOfPairs ()
