+++ Test/LeanCheck/Function.hs
@@ -1,0 +1,2 @@
+module Test.LeanCheck.Function () where
+import Test.LeanCheck.Function.ListsOfPairs ()
