+++ tests/test-fun.hs
@@ -1,12 +1,12 @@
 import System.Exit (exitFailure)
 import Data.List (elemIndices, sort, nub, delete)
 
-import Test.Check
-import Test.Check.Invariants
-import Test.Check.Function.ListsOfPairs
-import Test.Operators
-import Test.TypeBinding
-import Test.Types (Nat)
+import Test.LeanCheck
+import Test.LeanCheck.Invariants
+import Test.LeanCheck.Function.ListsOfPairs
+import Test.LeanCheck.Operators
+import Test.LeanCheck.TypeBinding
+import Test.LeanCheck.Types (Nat)
 
 
 main :: IO ()
