+++ tests/test-types.hs
@@ -1,7 +1,7 @@
 import System.Exit (exitFailure)
 import Data.List (elemIndices,delete)
-import Test.Types
-import Test.Check (list,fails)
+import Test.LeanCheck.Types
+import Test.LeanCheck (list,fails)
 
 main :: IO ()
 main =
